feat: タスクリストの notes 内 URL をタップ可能にする
リスト表示の notes 1行目を plain text で描画していたため、URL を含むタスクでも詳細パネルを開かないとリンクに飛べなかった。TaskDetail で既に使われていた URL→<a> 変換処理を src/lib/linkify.tsx に共通化し、TaskCard からも参照するように変更。

リンク click 時は e.stopPropagation でカードクリック（詳細パネル展開）の伝播を止める。target は _blank に統一して PC/モバイル両方で新規タブ遷移にした。
